WIP: Refactor `PoolsOverview`
